+++ src/main/kubectl.ts
@@ -9,7 +9,7 @@ import { helmCli } from "./helm/helm-cli"
 import { userStore } from "../common/user-store"
 import { customRequest } from "../common/request";
 import { getBundledKubectlVersion } from "../common/utils/app-version"
-import { isDevelopment, isWindows } from "../common/vars";
+import { isDevelopment, isWindows, isTestEnv } from "../common/vars";
 
 const bundledVersion = getBundledKubectlVersion()
 const kubectlMap: Map<string, string> = new Map([
@@ -22,9 +22,10 @@ const kubectlMap: Map<string, string> = new Map([
   ["1.13", "1.13.12"],
   ["1.14", "1.14.10"],
   ["1.15", "1.15.11"],
-  ["1.16", "1.16.8"],
+  ["1.16", "1.16.14"],
   ["1.17", bundledVersion],
-  ["1.18", "1.18.0"]
+  ["1.18", "1.18.8"],
+  ["1.19", "1.19.0"]
 ])
 
 const packageMirrors: Map<string, string> = new Map([
@@ -35,8 +36,9 @@ const packageMirrors: Map<string, string> = new Map([
 let bundledPath: string
 const initScriptVersionString = "# lens-initscript v3\n"
 
-if (isDevelopment) {
-  bundledPath = path.join(process.cwd(), "binaries", "client", process.platform, process.arch, "kubectl")
+if (isDevelopment || isTestEnv) {
+  const platformName = isWindows ? "windows" : process.platform
+  bundledPath = path.join(process.cwd(), "binaries", "client", platformName, process.arch, "kubectl")
 } else {
   bundledPath = path.join(process.resourcesPath, process.arch, "kubectl")
 }
@@ -58,6 +60,7 @@ export class Kubectl {
 
   public static readonly bundledKubectlPath = bundledPath
   public static readonly bundledKubectlVersion: string = bundledVersion
+  public static invalidBundle = false
   private static bundledInstance: Kubectl;
 
   // Returns the single bundled Kubectl instance
@@ -94,13 +97,38 @@ export class Kubectl {
 
     this.url = `${this.getDownloadMirror()}/v${this.kubectlVersion}/bin/${platformName}/${arch}/${binaryName}`
 
-    this.dirname = path.normalize(path.join(Kubectl.kubectlDir, this.kubectlVersion))
+    this.dirname = path.normalize(path.join(this.getDownloadDir(), this.kubectlVersion))
     this.path = path.join(this.dirname, binaryName)
   }
 
-  public async getPath(): Promise<string> {
+  public getBundledPath() {
+    return Kubectl.bundledKubectlPath
+  }
+
+  public getPathFromPreferences() {
+    return userStore.preferences?.kubectlBinariesPath || this.getBundledPath()
+  }
+
+  protected getDownloadDir() {
+    return userStore.preferences?.downloadBinariesPath || Kubectl.kubectlDir
+  }
+
+  public async getPath(bundled = false): Promise<string> {
+    if (userStore.preferences?.downloadKubectlBinaries === false) {
+      return this.getPathFromPreferences()
+    }
+
+    // return binary name if bundled path is not functional
+    if (!await this.checkBinary(this.getBundledPath(), false)) {
+      Kubectl.invalidBundle = true
+      return path.basename(bundledPath)
+    }
+
     try {
-      await this.ensureKubectl()
+      if (!await this.ensureKubectl()) {
+        logger.error("Failed to ensure kubectl, fallback to the bundled version")
+        return Kubectl.bundledKubectlPath
+      }
       return this.path
     } catch (err) {
       logger.error("Failed to ensure kubectl, fallback to the bundled version")
@@ -112,6 +140,7 @@ export class Kubectl {
   public async binDir() {
     try {
       await this.ensureKubectl()
+      await this.writeInitScripts()
       return this.dirname
     } catch (err) {
       logger.error(err)
@@ -119,16 +148,15 @@ export class Kubectl {
     }
   }
 
-  public async checkBinary(checkVersion = true) {
-    const exists = await pathExists(this.path)
+  public async checkBinary(path: string, checkVersion = true) {
+    const exists = await pathExists(path)
     if (exists) {
-      if (!checkVersion) {
-        return true
-      }
-
       try {
-        const { stdout } = await promiseExec(`"${this.path}" version --client=true -o json`)
+        const { stdout } = await promiseExec(`"${path}" version --client=true -o json`)
         const output = JSON.parse(stdout)
+        if (!checkVersion) {
+          return true
+        }
         let version: string = output.clientVersion.gitVersion
         if (version[0] === 'v') {
           version = version.slice(1)
@@ -165,20 +193,32 @@ export class Kubectl {
   }
 
   public async ensureKubectl(): Promise<boolean> {
+    if (userStore.preferences?.downloadKubectlBinaries === false) {
+      return true
+    }
+    if (Kubectl.invalidBundle) {
+      logger.error(`Detected invalid bundle binary, returning ...`)
+      return false
+    }
     await ensureDir(this.dirname, 0o755)
     return lockFile.lock(this.dirname).then(async (release) => {
       logger.debug(`Acquired a lock for ${this.kubectlVersion}`)
       const bundled = await this.checkBundled()
-      const isValid = await this.checkBinary(!bundled)
-      if (!isValid) {
+      let isValid = await this.checkBinary(this.path, !bundled)
+      if (!isValid && !bundled) {
         await this.downloadKubectl().catch((error) => {
           logger.error(error)
+          logger.debug(`Releasing lock for ${this.kubectlVersion}`)
+          release()
+          return false
         });
+        isValid = !await this.checkBinary(this.path, false)
       }
-      await this.writeInitScripts().catch((error) => {
-        logger.error("Failed to write init scripts");
-        logger.error(error)
-      })
+      if(!isValid) {
+        logger.debug(`Releasing lock for ${this.kubectlVersion}`)
+        release()
+        return false
+      }
       logger.debug(`Releasing lock for ${this.kubectlVersion}`)
       release()
       return true
@@ -221,71 +261,52 @@ export class Kubectl {
     })
   }
 
-  protected async scriptIsLatest(scriptPath: string) {
-    const scriptExists = await pathExists(scriptPath)
-    if (!scriptExists) return false
-
-    try {
-      const filehandle = await fs.promises.open(scriptPath, 'r')
-      const buffer = Buffer.alloc(40)
-      await filehandle.read(buffer, 0, 40, 0)
-      await filehandle.close()
-      return buffer.toString().startsWith(initScriptVersionString)
-    } catch (err) {
-      logger.error(err)
-      return false
-    }
-  }
-
   protected async writeInitScripts() {
+    const kubectlPath = userStore.preferences?.downloadKubectlBinaries ? this.dirname : path.dirname(this.getPathFromPreferences())
     const helmPath = helmCli.getBinaryDir()
     const fsPromises = fs.promises;
     const bashScriptPath = path.join(this.dirname, '.bash_set_path')
-    const bashScriptIsLatest = await this.scriptIsLatest(bashScriptPath)
-    if (!bashScriptIsLatest) {
-      let bashScript = "" + initScriptVersionString
-      bashScript += "tempkubeconfig=\"$KUBECONFIG\"\n"
-      bashScript += "test -f \"/etc/profile\" && . \"/etc/profile\"\n"
-      bashScript += "if test -f \"$HOME/.bash_profile\"; then\n"
-      bashScript += "  . \"$HOME/.bash_profile\"\n"
-      bashScript += "elif test -f \"$HOME/.bash_login\"; then\n"
-      bashScript += "  . \"$HOME/.bash_login\"\n"
-      bashScript += "elif test -f \"$HOME/.profile\"; then\n"
-      bashScript += "  . \"$HOME/.profile\"\n"
-      bashScript += "fi\n"
-      bashScript += `export PATH="${this.dirname}:${helmPath}:$PATH"\n`
-      bashScript += "export KUBECONFIG=\"$tempkubeconfig\"\n"
-      bashScript += "unset tempkubeconfig\n"
-      await fsPromises.writeFile(bashScriptPath, bashScript.toString(), { mode: 0o644 })
-    }
+
+    let bashScript = "" + initScriptVersionString
+    bashScript += "tempkubeconfig=\"$KUBECONFIG\"\n"
+    bashScript += "test -f \"/etc/profile\" && . \"/etc/profile\"\n"
+    bashScript += "if test -f \"$HOME/.bash_profile\"; then\n"
+    bashScript += "  . \"$HOME/.bash_profile\"\n"
+    bashScript += "elif test -f \"$HOME/.bash_login\"; then\n"
+    bashScript += "  . \"$HOME/.bash_login\"\n"
+    bashScript += "elif test -f \"$HOME/.profile\"; then\n"
+    bashScript += "  . \"$HOME/.profile\"\n"
+    bashScript += "fi\n"
+    bashScript += `export PATH="${helmPath}:${kubectlPath}:$PATH"\n`
+    bashScript += "export KUBECONFIG=\"$tempkubeconfig\"\n"
+    bashScript += "unset tempkubeconfig\n"
+    await fsPromises.writeFile(bashScriptPath, bashScript.toString(), { mode: 0o644 })
 
     const zshScriptPath = path.join(this.dirname, '.zlogin')
-    const zshScriptIsLatest = await this.scriptIsLatest(zshScriptPath)
-    if (!zshScriptIsLatest) {
-      let zshScript = "" + initScriptVersionString
 
-      zshScript += "tempkubeconfig=\"$KUBECONFIG\"\n"
-      // restore previous ZDOTDIR
-      zshScript += "export ZDOTDIR=\"$OLD_ZDOTDIR\"\n"
-      // source all the files
-      zshScript += "test -f \"$OLD_ZDOTDIR/.zshenv\" && . \"$OLD_ZDOTDIR/.zshenv\"\n"
-      zshScript += "test -f \"$OLD_ZDOTDIR/.zprofile\" && . \"$OLD_ZDOTDIR/.zprofile\"\n"
-      zshScript += "test -f \"$OLD_ZDOTDIR/.zlogin\" && . \"$OLD_ZDOTDIR/.zlogin\"\n"
-      zshScript += "test -f \"$OLD_ZDOTDIR/.zshrc\" && . \"$OLD_ZDOTDIR/.zshrc\"\n"
+    let zshScript = "" + initScriptVersionString
 
-      // voodoo to replace any previous occurrences of kubectl path in the PATH
-      zshScript += `kubectlpath=\"${this.dirname}"\n`
-      zshScript += `helmpath=\"${helmPath}"\n`
-      zshScript += "p=\":$kubectlpath:\"\n"
-      zshScript += "d=\":$PATH:\"\n"
-      zshScript += "d=${d//$p/:}\n"
-      zshScript += "d=${d/#:/}\n"
-      zshScript += "export PATH=\"$kubectlpath:$helmpath:${d/%:/}\"\n"
-      zshScript += "export KUBECONFIG=\"$tempkubeconfig\"\n"
-      zshScript += "unset tempkubeconfig\n"
-      zshScript += "unset OLD_ZDOTDIR\n"
-      await fsPromises.writeFile(zshScriptPath, zshScript.toString(), { mode: 0o644 })
-    }
+    zshScript += "tempkubeconfig=\"$KUBECONFIG\"\n"
+    // restore previous ZDOTDIR
+    zshScript += "export ZDOTDIR=\"$OLD_ZDOTDIR\"\n"
+    // source all the files
+    zshScript += "test -f \"$OLD_ZDOTDIR/.zshenv\" && . \"$OLD_ZDOTDIR/.zshenv\"\n"
+    zshScript += "test -f \"$OLD_ZDOTDIR/.zprofile\" && . \"$OLD_ZDOTDIR/.zprofile\"\n"
+    zshScript += "test -f \"$OLD_ZDOTDIR/.zlogin\" && . \"$OLD_ZDOTDIR/.zlogin\"\n"
+    zshScript += "test -f \"$OLD_ZDOTDIR/.zshrc\" && . \"$OLD_ZDOTDIR/.zshrc\"\n"
+
+    // voodoo to replace any previous occurrences of kubectl path in the PATH
+    zshScript += `kubectlpath=\"${kubectlPath}"\n`
+    zshScript += `helmpath=\"${helmPath}"\n`
+    zshScript += "p=\":$kubectlpath:\"\n"
+    zshScript += "d=\":$PATH:\"\n"
+    zshScript += "d=${d//$p/:}\n"
+    zshScript += "d=${d/#:/}\n"
+    zshScript += "export PATH=\"$helmpath:$kubectlpath:${d/%:/}\"\n"
+    zshScript += "export KUBECONFIG=\"$tempkubeconfig\"\n"
+    zshScript += "unset tempkubeconfig\n"
+    zshScript += "unset OLD_ZDOTDIR\n"
+    await fsPromises.writeFile(zshScriptPath, zshScript.toString(), { mode: 0o644 })
   }
 
   protected getDownloadMirror() {