+++ src/main/menu.ts
@@ -43,7 +43,7 @@ export function buildMenu(windowManager: WindowManager) {
       `${appName}: ${app.getVersion()}`,
       `Electron: ${process.versions.electron}`,
       `Chrome: ${process.versions.chrome}`,
-      `Copyright 2020 Copyright 2020 Mirantis, Inc.`,
+      `Copyright 2020 Mirantis, Inc.`,
     ]
     dialog.showMessageBoxSync(browserWindow, {
       title: `${isWindows ? " ".repeat(2) : ""}${appName}`,
@@ -163,14 +163,14 @@ export function buildMenu(windowManager: WindowManager) {
         label: 'Forward',
         accelerator: 'CmdOrCtrl+]',
         click() {
-          webContents.getFocusedWebContents()?.goForward();
+          webContents.getFocusedWebContents()?.goForward()
         }
       },
       {
         label: 'Reload',
         accelerator: 'CmdOrCtrl+R',
         click() {
-          webContents.getFocusedWebContents()?.reload();
+          windowManager.reload({ channel: "menu:reload" });
         }
       },
       { role: 'toggleDevTools' },